Implement < as well as isless
but didn't likewise move `<`. Because `<` falls back to `isless`,
this meant that comparisons like `Gray(0.3) < Gray(NaN)`
returned a different answer from `0.3 < NaN`.

This PR also moves code and tests from "traits" to "operations."
