+++ src/ColorTypes.jl
@@ -6,7 +6,8 @@ using Base: @pure
 
 const Fractional = Union{AbstractFloat, FixedPoint}
 
-import Base: ==, isapprox, hash, convert, eltype, length, show, oneunit, zero, reinterpret, getindex
+import Base: ==, <, isless, isapprox, hash, convert, eltype, length, show,
+             oneunit, zero, reinterpret, getindex
 using Random
 import Random: rand
